Image Similarity Function Enhanced

+++ server/main.py
@@ -1,16 +1,20 @@
+from dotenv import load_dotenv
 from fastapi import FastAPI
-import clip
+from openai import OpenAI
+import os
 import pandas as pd
-from PIL import Image
-import torch
 
 from server.routes.keywords import router as keywords_router
-from server.routes.image import router as image_router
+from server.routes.image import router as image_router, get_image_embedding
+
+load_dotenv()
 
 app = FastAPI()
 
 app.include_router(image_router, prefix="/image")
 app.include_router(keywords_router, prefix="/keywords")
+
+client = OpenAI()
 
 
 @app.get("/")
@@ -19,24 +23,14 @@ def hello():
 
 
 @app.on_event("startup")
-def get_features():
-    df = pd.read_csv("***REMOVED***/characters.csv", index_col=False)
+def generate_pickle():
+    if os.path.exists("***REMOVED***/characters.pkl"):
+        return
+    df = pd.read_excel("***REMOVED***/characters.xlsx")
     for column in df.columns:
         df[column] = df[column].apply(lambda x: str(x).replace(",", ""))
     df["id"] = range(len(df))
     df.index = df["id"]
 
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-    model, preprocess = clip.load("ViT-B/32", device=device)
-
-    df["image"] = df["id"].apply(lambda x: Image.open(f"***REMOVED***/images/{x}.png"))
-
-    def encode_image(image):
-        image_input = preprocess(image).unsqueeze(0).to(device)
-        with torch.no_grad():
-            image_features = model.encode_image(image_input)
-        return image_features.cpu().numpy().tolist()
-
-    df["embedding"] = df["image"].apply(encode_image)
-
+    df["embedding"] = df["description"].apply(get_image_embedding)
     df.to_pickle("***REMOVED***/characters.pkl")
